Update README.md
Im adding the survey paper "A novel instrument to compare dynamic object detection algorithms"

A briefly resume:
Nowadays, the amount of dynamic object detection in video sequences algorithms has increased considerably. Notwithstanding the many efforts to provide benchmarking resource, a standard methodology to achieve this evaluation does not exist. Most of the existing benchmarking resources concentrate on the evaluation of the algorithms from a rigid perspective by using just quantitative metric values of the performance. However, these evaluations do not consider important criteria like documentation, auto-adaptability, novelty, speed, which are important factors to consider from a scientific and/or real word application. Therefore, this paper proposes a new methodology to evaluate, compare, and select dynamic object detection algorithms by considering the criteria previously mentioned including performance. The new methodology was developed by analyzing 119 algorithms and the databases CDnet2014, CDnet2012 and BMC. The findings indicate that the proposed methodology preserves consistence with some of the rankings in the databases, but it also provides more complete and useful information in the evaluation of the algorithm.